Precision improved and slight refactor

+++ Verlet.cs
@@ -12,7 +12,7 @@ namespace VerletChainAlgo
         public bool Active = true;
 
         public int segmentCount;
-        public int segmentDistance;
+        public float segmentDistance;
 
         public int constraintRepetitions = 2;
         public float drag = 1f;
@@ -26,7 +26,7 @@ namespace VerletChainAlgo
 
         //advanced variables
         public bool customDistances = false;
-        public List<int> segmentDistances;//length must match the segment count
+        public List<float> segmentDistances;//length must match the segment count
 
         public bool customGravity = false;
         public List<Vector2> forceGravities;//length must match the segment count
@@ -35,7 +35,7 @@ namespace VerletChainAlgo
 
         public List<RopeSegment> ropeSegments;
 
-        public VerletChainInstance(int SegCount, bool specialDraw, int SegDistance)
+        public VerletChainInstance(int SegCount, bool specialDraw, float SegDistance)
         {
             segmentCount = SegCount;
             segmentDistance = SegDistance;
@@ -43,7 +43,7 @@ namespace VerletChainAlgo
             Start();
         }
 
-        public VerletChainInstance(int SegCount, Vector2? StartPoint = null, Vector2? EndPoint = null, int SegDistance = 5, Vector2? Grav = null)
+        public VerletChainInstance(int SegCount, Vector2? StartPoint = null, Vector2? EndPoint = null, float SegDistance = 5, Vector2? Grav = null)
         {
             segmentCount = SegCount;
             segmentDistance = SegDistance;
@@ -57,9 +57,9 @@ namespace VerletChainAlgo
             Start(EndPoint != null);
         }
 
-        public VerletChainInstance(int SegCount, Vector2? StartPoint = null, Vector2? EndPoint = null, int SegDistance = 5, Vector2? Grav = null,
+        public VerletChainInstance(int SegCount, Vector2? StartPoint = null, Vector2? EndPoint = null, float SegDistance = 5, Vector2? Grav = null,
             bool CustomGravs = false, List<Vector2> SegGravs = null,
-            bool CustomDists = false, List<int> SegDists = null)
+            bool CustomDists = false, List<float> SegDists = null)
         {
             segmentCount = SegCount;
             segmentDistance = SegDistance;
@@ -96,7 +96,7 @@ namespace VerletChainAlgo
                 }
                 else
                 {
-                    int distance = customDistances ? segmentDistances[i] : segmentDistance;
+                    float distance = customDistances ? segmentDistances[i] : segmentDistance;
                     Vector2 spawnGrav = customGravity ? forceGravities[i] * forceGravity : forceGravity;
                     if (spawnGrav != Vector2.Zero)
                         nextRopePoint += Vector2.Normalize(spawnGrav) * distance;
